pyproject.toml
[tool.black]
line-length = 120

[tool.poetry]
name = "yaml-patch"
version = "0.2.0"
description = "Patch yaml strings"
readme = "README.md"
authors = ["Diogo de Campos <campos.ddc@gmail.com>"]
homepage = "https://github.com/campos-ddc/yaml-patch"
license = "MIT"

[tool.poetry.scripts]
yaml-patch = 'yaml_patch.cli:cli'

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.0.1"
"ruamel.yaml" = "^0.17.11"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.7b0"
tox = "^3.24.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
